functions/requirements.txt
firebase-functions
firebase-admin
google-generativeai
pydantic
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
requests